Extend hotels api to accept a search term
Named the search term `freetext-search` to make it as clear as possible that the value is used for search, and that
it's a free-text search rather than a facet filtering.

I think the name gives room for other types of search, eg a location search where it's only searching by location.