Extend API contract to support returning country and city results
Changing the structure to prepare for implementation of the api logic.  There's a lot of changes to the API logic, the
UI state, and the UI presentation of the result

I moved the hotel type to the interfaces folder and made it an interface for consistency.  I didn't map all hotel fields
since they are not all used.

At this stage, since the API isn't returning city/country results, it doesn't matter the props arestring arrays because in
the response the cities and countries will be an empty array

Duplicating the interfaces in the api and client projects because I'm assuming the FE and the BE are in different repos.

Removed `fetchAndFilterHotels` because it's not doing much now that the filtering logic is removed