Extend /hotels API to return matching cities and countries
Depending on how the api is currently used it may be best to create a new endpoint.  For the purposes
of the tech test it seemed that I can assume the client is the only consumer of the API, and therefore can
make a breaking change.

Next step: update the UI to show matching cities and countries