Move mongodb client init logic to central function
With the soon-to-be-added endpoints for retrieving a single hotel/city/country, the
mongodb client will be used more widely.  Rather than duplicate the logic in the new
endpoints I decided to split the logic into a new function

Seeding only needs to be run once, so that's outside the client initialisation.

I only initialise the client if it's not already initialised.  This shpould make the site
slightly more performant across searches and different pages